# schema/openapi.yaml
...
openapi: 3.1.0
info:
  title: slim-m API
  version: 0.5.0
  description: >
    The server surface: liveness, auth with opaque rotating tokens, channel
    management, text messaging with reactions, polls, deletion, and full-text
    search, user profiles and the member list, devices, blocking, report
    intake and triage, invite codes, read state, bundled catch-up sync,
    account deletion, admin-issued password reset, role management and
    channel permission overwrites, presence and typing indicators, and the
    WebSocket event envelope.

    Identity and ordering are separate concerns. A message's identity is a
    client-generated UUIDv7, which also makes sending idempotent; its order is a
    per-channel monotonic `seq` assigned by the server, which is the
    authoritative cursor for pagination and sync.

    Authentication is a bearer access token. Access tokens are short-lived and
    exchanged at `/auth/refresh`, which rotates the refresh token. Opening a
    WebSocket uses a single-use ticket from `/auth/ws-ticket`, because an
    upgrade request cannot reliably carry an Authorization header.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      operationId: refresh
      tags: [auth]
      summary: Exchange a refresh token for a new pair.
      description: >
        Rotation is single-use: the presented token is spent and a new pair
        issued. Replaying a long-spent token is treated as a leak and revokes
        the whole family and its session. A replay moments after a successful
        rotation is treated as the client racing itself and is denied without
        revoking anything.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          $ref: "#/components/responses/Tokens"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /ws:
    get:
      operationId: connectWebSocket
      tags: [system]
      summary: Open the event WebSocket.
      security: []
      description: >
        Upgrade, then send a `hello` frame carrying a ticket from
        `/auth/ws-ticket` and the protocol version. The server redeems the
        ticket and replies with its own `hello`. After that the connection
        carries server-to-client events, authorized per event, plus a ping and
        pong keepalive. See the ClientFrame and ServerFrame schemas.
      responses:
        "101":
          description: Switching protocols.
        "503":
          description: The connection ceiling was reached.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An opaque access token from register, login, or refresh.
...
  responses:
    Tokens:
      description: A new access and refresh token pair.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TokenPair"
...
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: The caller is over its rate budget for this traffic class.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A short, stable, human-readable reason.
...
    TokenPair:
      type: object
      required: [user_id, access_token, refresh_token, access_expires_at]
      properties:
        user_id:
          type: string
          format: uuid
        access_token:
          type: string
          description: Bearer token for authenticated requests. Short-lived.
        refresh_token:
          type: string
          description: Single-use token that rotates on each refresh.
        access_expires_at:
          type: integer
          format: int64
          description: Unix milliseconds at which the access token expires.
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...